feat: return the topmost overlay if present on tab overlay
`QuickSelect` or `CopyMode` overlays have same ID as the pane on which it is assigned to.
`DisplayText`, `CloseCurrentPane`, `PromptInputLine`, `InputSelector`, `Confirmation` overlays result in new pane (`TermWizTerminalPane`) getting added to the mux.

Based on above points, the pane structure would be:

1. QuickSelect (QuickSelectOverlay) on LocalPane
```
LocalPane (p1)
QuickSelect (p1) (overlay of p1)
```

2. DisplayText (TermWizTerminalPane) on LocalPane
```
DisplayText (p1) (overlay of currently active tab)
```

3. QuickSelect (QuickSelectOverlay) on DisplayText (TermWizTerminalPane)
```
DisplayText (p1) (overlay of currently active tab)
QuickSelect (p1) (overlay of p1)
```

The part of the region of the pane or tab to be rendered is fetched using `get_active_pane_or_overlay`
Cases 1 and 2 would work as expected with the earlier implementation of `get_active_pane_or_overlay`
But case 3 would fail, because even though QuickSelect overlay is assigned to DisplayText overlay, the `get_active_pane_or_overlay` used to return only the first overlay if it was active.

This change updates `get_active_pane_or_overlay` method to returns the topmost overlay in the pane region if present