fix: tmux -CC mode not parsing output correctly
refs: #6769

The tmux output may split into multiple parts, if this spliting happens
in the middle of multiple bytes UTF8 characters, then the issue raised.
That the first part will have incomplete UTF8 character while the second
part will have a invalid UTF8 character.

In this case, we have to treat the output text as bytes rather than
UTF8 string.

This fix unvis the output text and send to pane as bytes.