Remove `mode` from `PushKittyState` CSI representation
The [Kitty Keyboard document] specifies

    CSI = flags ; mode u

in the "Progressive enhancement" section stating that the `mode` (which
controls how the flags are set) is optional. It then talks about pushing
and popping flags - useful for applications using an alternate screen -
with the sequence

    CSI > flags u  # for push, if flags omitted default to zero
    CSI < number u # to pop number entries, defaulting to 1 if unspecified

Termwiz previously also emitted and accepted the `mode` for the "push"
command, but I believe it is a mistake. The docs don't explicitly
mention the mode here - which you might think is because it is
optional - but also the Kitty codebase does not accept a mode:

* [`screen_push_key_encoding_flags`] does not accept a `how` parameter
  like the definition of `screen_set_key_encoding_flags` above it.
* the Python client used to record and replay commands has a function of
  the same name which also lacks a `how` parameter:
  <https://github.com/kovidgoyal/kitty/blob/b0371d970fac15a414ea15fcf25c20269b4b4f65/kitty/client.py#L184-L185>

[Kitty Keyboard document]: https://sw.kovidgoyal.net/kitty/keyboard-protocol/
[`screen_push_key_encoding_flags`]: https://github.com/kovidgoyal/kitty/blob/b0371d970fac15a414ea15fcf25c20269b4b4f65/kitty/screen.c#L1692-L1704