feat: store 2 optional overlays per tab
This is in order to not clear the DisplayText overlay in case any other
overlay is activated. DisplayText overlay is activated at the bottom
layer (layer 0), and the other overlays are activated at the top layer
(layer 1). Layer 1 overlay can be activated irrespective of whether
layer 0 overlay is active. Layer 0 overlay can be activated only when
layer 1 overlay is not active