refactor: add method to cancel all overlays for a tab
Ideally we could call cancel_overlay_for_tab for each layer.
But it would lead to borrow error because of the below reason.
Because the cancel_overlay_for_tab cancels overlay by layer, we perform
the below steps:
* To obtain layers, we call tab_state method which takes an immutable
  reference to self
* We pass the layer to cancel_overlay_for_tab method which takes a
  mutable reference to self.
So we define a separate method to cancel all overlays for a tab.