feat: add allow_nil field to EditCommandOption struct
If allow_nil field is set to false, then user is always prompted for
value whenever the flag is triggered. If set to true, the user is
only prompted for value if the option was previously unset while
triggering the flag, else it sets the option to false. If not specified,
the field is set to true