fix: increase row overhead to 4 to avoid crossing the viewport
Increase row overhead to 4 in order to avoid crossing the viewport when
the selector is activated. Previously when the no. of choices was
greater or equal to the difference between no. of rows and row overhead,
the viewport was being crossed because of the selector reaching the
first row. There are at least 4 rows that are not reserved for the
entries. The topmost row, the bottommost row, row for description of the
selector, row for separator between content and selector