feat: add optional field context to display in TransientMenu
Additional information can now be rendered in TransientMenu. This is
useful in situations when chaining TransientMenus. While the state can
be passed from one TransientMenu action to a different TransientMenu, it
is helpful to display the state in the new TransientMenu to help keep
track of our selected choices and actions performed.
A new overlay will be added that displays a buffer containing the
choices we passed and list of key action mappings which can be composed
with TransientMenu. Displaying the context is also useful in this case.