fix: update length of line to be drawn below description
The length of line to be drawn below description in TransientMenu should
be based on the length of the Line obtained on calling
crate::tabbar::parse_status_text method. This is because if we apply
formatting to the string, the length of the string is not the same as
the no. of characters present in the string