fix: reset traversed_nodes if end_of_word reached with empty choices
If we trigger an argument without selecting any choice, and then select
choices and trigger an argument, it would fail because the
traversed_nodes vector is already populated and a keypress would be
checked against entries in the leaf node instead of the root node. So we
need to reset traversed_nodes to root_node in order to fix this.