docs: Improve get_progress example circle rendering
This improves the circle rendering in the get_progress example, so that
the circle shows *completed* progress only. Meaning, at 12 or lower it
shows an empty circle, at 13 it shows the first 1/8 slice. The full
circle is only shown at 100.

Additionally this adds an indeterminate circle and also improves the
spacing around the progress display.