Abandon kitty's legacy key event encoding
WezTerm incorrectly implements kitty's Legacy key event encoding:

https://sw.kovidgoyal.net/kitty/keyboard-protocol/#legacy-key-event-encoding

This section is not meant to be implemented as a protocol that can be
activated. It merely describes the default plain behavior of kitty,
which is simply general behavior that most terminals implement very
similarly and which must conform to terminfo. As a consequence, there is
no mechanism to turn this mode on or off in the kitty specification.

Therefore, WezTerm should not implement the legacy behavior of kitty at
all, because its default behavior is already fully sufficient for that
purpose. WezTerm (and other terminals) should implement only Progressive
enhancements, which are fully switchable. Setting the progressive
enhancement flags to 0 should disable the kitty protocol completely.
Currently, setting the flags to 0 only switches WezTerm into this "kitty
legacy mode", which can't be disabled in any other way than by sending
the pop (CSI < u) sequence repeatedly until there is no pushed state
left in the kitty enhancement flags stack.

This bug is reproducible, for example, when using vim: enable
config.enable_kitty_keyboard, launch vim, and then exit vim.  After
that, the arrow keys will not work in bash, producing incompatible kitty
sequences. (To make it even worse, unlike kitty itself, WezTerm produces
incompatible legacy sequences because it incorrectly sends Num Lock and
Caps Lock modifier codes, which are not allowed in legacy mode, and it
also sends modifier code even when no modifier is set, which also
conflicts with kitty's legacy mode.)

This commit changes the behavior to treat progressive enhancement flags
set to 0 as meaning the kitty is completely disabled, which is the
correct behavior.

Closes: #3593
Closes: #6434

Signed-off-by: Jakub Horký <jakub.git@horky.net>