Convert event handling to select
`smol::or` is a awkward because it requires all selected events to
resolve to the same type. This leads to some not particularly elegant or
correct type gymnastics.

The previous code waited for the underlying socket to become readable,
and then used that to trigger an atomic decode of a PDU. However this
could actually deadlock because the socket might only be readable
because of protocol-level traffic, not because user PDU traffic is about
to arrive.

The deadlock occurs on QUIC server connections, because the server can't
accept the connection until the client actually sends something (the
initial PDU), but often the client would get stuck trying to read a PDU
from the server which can't happen.

The "obvious" fix of selecting on the PDU itself results in cancel-safety
problems, as `select` will cancel a partially read PDU if one of its
other arms has activity, so we need to rely on StreamPeek<S>::peek to
wait for a PDU, then read it without interleaving it with other events.

In theory this is also an issue with TLS, but it seems in practice that
a readable socket does always result in user-level PDU traffic so it
seems to work. But the new code is strictly more correct.