Add StreamPeek wrapper to codec for cancel-safe PDU reading
This allows a single byte to be peeked from an underlying stream via
`async fn peek(&mut self) -> Result<u8>`.

The intended use is so that when using `futures::select!` you can select
on a TLS or QUIC stream waiting for the next user-level byte, then
process a whole PDU as a single unit.

The current implementation relies on waiting for the underlying socket
to be readable, but this could just be protocol-level traffic, but that
doesn't mean there will necessarily user-level data any time soon.

For TLS this seems to work out OK in practice, but for QUIC a server
can't complete a connection until the client has sent something,
so waiting for user data rather than protocol traffic will avoid a
deadlock.