docs: linuxbrew install command specify formula
`brew install wezterm` attempts to install the macOS package and fails with `Error: macOS is required for this software.` because it treats `wezterm` as a cask, as per the `Warning: Treating wezterm as a cask. For the formula, use wezterm/wezterm-linuxbrew/wezterm or specify the `--formula` flag. To silence this message, use the `--cask` flag.`